Move serde test helpers to an appropriate module
