+++ src/naive/datetime.rs
@@ -1718,10 +1718,12 @@ pub(super) mod rustc_serialize {
 
 /// Tools to help serializing/deserializing `NaiveDateTime`s
 #[cfg(feature = "serde")]
-pub(super) mod serde {
-    use super::NaiveDateTime;
+pub(crate) mod serde {
     use core::fmt;
     use serde::{de, ser};
+
+    use super::NaiveDateTime;
+    use crate::offset::LocalResult;
 
     /// Serialize a `NaiveDateTime` as an RFC 3339 string
     ///
@@ -1811,7 +1813,8 @@ pub(super) mod serde {
         use core::fmt;
         use serde::{de, ser};
 
-        use crate::{ne_timestamp, NaiveDateTime};
+        use super::ne_timestamp;
+        use crate::NaiveDateTime;
 
         /// Serialize a UTC datetime into an integer number of nanoseconds since the epoch
         ///
@@ -1963,7 +1966,8 @@ pub(super) mod serde {
         use core::fmt;
         use serde::{de, ser};
 
-        use crate::{ne_timestamp, NaiveDateTime};
+        use super::ne_timestamp;
+        use crate::NaiveDateTime;
 
         /// Serialize a UTC datetime into an integer number of milliseconds since the epoch
         ///
@@ -2112,7 +2116,8 @@ pub(super) mod serde {
         use core::fmt;
         use serde::{de, ser};
 
-        use crate::{ne_timestamp, NaiveDateTime};
+        use super::ne_timestamp;
+        use crate::NaiveDateTime;
 
         /// Serialize a UTC datetime into an integer number of seconds since the epoch
         ///
@@ -2261,6 +2266,58 @@ pub(super) mod serde {
         let actual = deserialize::<Test>(&(bytes)).unwrap();
 
         assert_eq!(expected, actual);
+    }
+
+    // lik? function to convert a LocalResult into a serde-ish Result
+    pub(crate) fn serde_from<T, E, V>(me: LocalResult<T>, ts: &V) -> Result<T, E>
+    where
+        E: de::Error,
+        V: fmt::Display,
+        T: fmt::Display,
+    {
+        match me {
+            LocalResult::None => Err(E::custom(ne_timestamp(ts))),
+            LocalResult::Ambiguous(min, max) => {
+                Err(E::custom(SerdeError::Ambiguous { timestamp: ts, min: min, max: max }))
+            }
+            LocalResult::Single(val) => Ok(val),
+        }
+    }
+
+    #[cfg(feature = "serde")]
+    enum SerdeError<V: fmt::Display, D: fmt::Display> {
+        NonExistent { timestamp: V },
+        Ambiguous { timestamp: V, min: D, max: D },
+    }
+
+    /// Construct a [`SerdeError::NonExistent`]
+    #[cfg(feature = "serde")]
+    fn ne_timestamp<T: fmt::Display>(ts: T) -> SerdeError<T, u8> {
+        SerdeError::NonExistent::<T, u8> { timestamp: ts }
+    }
+
+    #[cfg(feature = "serde")]
+    impl<V: fmt::Display, D: fmt::Display> fmt::Debug for SerdeError<V, D> {
+        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+            write!(f, "ChronoSerdeError({})", self)
+        }
+    }
+
+    // impl<V: fmt::Display, D: fmt::Debug> core::error::Error for SerdeError<V, D> {}
+    #[cfg(feature = "serde")]
+    impl<V: fmt::Display, D: fmt::Display> fmt::Display for SerdeError<V, D> {
+        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+            match self {
+                &SerdeError::NonExistent { ref timestamp } => {
+                    write!(f, "value is not a legal timestamp: {}", timestamp)
+                }
+                &SerdeError::Ambiguous { ref timestamp, ref min, ref max } => write!(
+                    f,
+                    "value is an ambiguous timestamp: {}, could be either of {}, {}",
+                    timestamp, min, max
+                ),
+            }
+        }
     }
 }
